+++ src/components/Contestant.svelte
